api/requirements.txt
alembic==1.11.1
bcrypt==4.0.1
blinker==1.6.2
boto3==1.26.142
botocore==1.29.142
click==8.1.3
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
protobuf==4.21.12
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
urllib3==1.26.16
Werkzeug==2.3.4
